feat: add cursor support
